+++ src/main/kotlin/com/wafflestudio/toyproject/team4/core/purchase/api/request/PurchasesRequest.kt
@@ -1,4 +1,4 @@
-package com.wafflestudio.toyproject.team4.core.user.api.request
+package com.wafflestudio.toyproject.team4.core.purchase.api.request
 
 data class PurchasesRequest(
     val purchaseItems: List<PurchaseRequest>
