+++ src/main/kotlin/com/wafflestudio/toyproject/team4/core/purchase/api/response/PurchaseItemsResponse.kt
@@ -1,0 +1,7 @@
+package com.wafflestudio.toyproject.team4.core.purchase.api.response
+
+import com.wafflestudio.toyproject.team4.core.purchase.domain.Purchase
+
+data class PurchaseItemsResponse(
+    val purchaseItems: List<Purchase>
+)
