+++ src/main/kotlin/com/wafflestudio/toyproject/team4/core/purchase/domain/Purchase.kt
@@ -1,7 +1,8 @@
-package com.wafflestudio.toyproject.team4.core.user.domain
+package com.wafflestudio.toyproject.team4.core.purchase.domain
 
 import com.wafflestudio.toyproject.team4.core.item.domain.Item
-import com.wafflestudio.toyproject.team4.core.user.database.PurchaseEntity
+import com.wafflestudio.toyproject.team4.core.purchase.database.PurchaseEntity
+import com.wafflestudio.toyproject.team4.core.user.domain.User
 import java.time.LocalDateTime
 
 data class Purchase(
