+++ src/main/kotlin/com/wafflestudio/toyproject/team4/core/user/api/UserController.kt
@@ -3,9 +3,6 @@ package com.wafflestudio.toyproject.team4.core.user.api
 import com.wafflestudio.toyproject.team4.common.Authenticated
 import com.wafflestudio.toyproject.team4.common.UserContext
 import com.wafflestudio.toyproject.team4.core.user.api.request.PatchMeRequest
-import com.wafflestudio.toyproject.team4.core.user.api.request.PatchShoppingCartRequest
-import com.wafflestudio.toyproject.team4.core.user.api.request.PostShoppingCartRequest
-import com.wafflestudio.toyproject.team4.core.user.api.request.PurchasesRequest
 import com.wafflestudio.toyproject.team4.core.user.api.request.RecentlyViewedRequest
 import com.wafflestudio.toyproject.team4.core.user.api.response.UserResponse
 import com.wafflestudio.toyproject.team4.core.user.service.AuthTokenService
@@ -81,54 +78,6 @@ class UserController(
         @UserContext username: String,
         @RequestBody patchMeRequest: PatchMeRequest
     ) = ResponseEntity(userService.patchMe(username, patchMeRequest), HttpStatus.OK)
-
-    @Authenticated
-    @GetMapping("/me/purchases")
-    fun getPurchases(
-        @RequestHeader(value = "Authorization") authorization: String,
-        @UserContext username: String,
-    ) = ResponseEntity(userService.getPurchases(username), HttpStatus.OK)
-
-    @Authenticated
-    @PostMapping("/me/purchases")
-    fun postPurchases(
-        @RequestBody request: PurchasesRequest,
-        @RequestHeader(value = "Authorization") authorization: String,
-        @UserContext username: String,
-    ) {
-        userService.postPurchases(username, request)
-    }
-
-    @Authenticated
-    @GetMapping("/me/shopping-cart")
-    fun getShoppingCart(
-        @RequestHeader(value = "Authorization") authorization: String,
-        @UserContext username: String,
-    ) = ResponseEntity(userService.getShoppingCart(username), HttpStatus.OK)
-
-    @Authenticated
-    @PostMapping("/me/shopping-cart")
-    fun postShoppingCart(
-        @RequestHeader(value = "Authorization") authorization: String,
-        @UserContext username: String,
-        @RequestBody postShoppingCartRequest: PostShoppingCartRequest
-    ) = ResponseEntity(userService.postShoppingCart(username, postShoppingCartRequest), HttpStatus.CREATED)
-
-    @Authenticated
-    @PatchMapping("/me/shopping-cart")
-    fun patchShoppingCart(
-        @RequestHeader(value = "Authorization") authorization: String,
-        @UserContext username: String,
-        @RequestBody patchShoppingCartRequest: PatchShoppingCartRequest
-    ) = ResponseEntity(userService.patchShoppingCart(username, patchShoppingCartRequest), HttpStatus.OK)
-
-    @Authenticated
-    @DeleteMapping("/me/shopping-cart/{id}")
-    fun deleteShoppingCart(
-        @RequestHeader(value = "Authorization") authorization: String,
-        @UserContext username: String,
-        @PathVariable(value = "id") cartItemId: Long
-    ) = ResponseEntity(userService.deleteShoppingCart(username, cartItemId), HttpStatus.OK)
 
     @Authenticated
     @GetMapping("/me/recently-viewed")
