--- src/main/kotlin/com/wafflestudio/toyproject/team4/core/user/api/response/CartItemsResponse.kt
@@ -1,7 +1,0 @@
-package com.wafflestudio.toyproject.team4.core.user.api.response
-
-import com.wafflestudio.toyproject.team4.core.user.domain.CartItem
-
-data class CartItemsResponse(
-    val cartItems: List<CartItem>
-)
