--- src/main/kotlin/com/wafflestudio/toyproject/team4/core/user/api/response/PurchaseItemsResponse.kt
@@ -1,7 +1,0 @@
-package com.wafflestudio.toyproject.team4.core.user.api.response
-
-import com.wafflestudio.toyproject.team4.core.user.domain.Purchase
-
-data class PurchaseItemsResponse(
-    val purchaseItems: List<Purchase>
-)
